requirements-dev.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.13
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncssh==2.14.1
atpublic==4.0
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.10.1
bleach==6.1.0
blis==0.7.11
bump2version==1.0.1
cachetools==5.3.2
catalogue==2.0.10
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.3
configobj==5.0.8
contourpy==1.1.1
coverage==7.3.2
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.7
distro==1.8.0
docutils==0.18.1
dpath==2.1.6
dulwich==0.21.6
dvc==3.30.1
dvc-data==2.22.0
dvc-gdrive==2.20.0
dvc-http==2.30.2
dvc-objects==1.2.0
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
entrypoints==0.4
# Editable Git install with no remote (eta==0.0.1.dev0)
-e /Users/roman/PycharmProjects/ETA
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
grandalf==0.8
gto==1.5.0
httplib2==0.22.0
huggingface-hub==0.17.3
hydra-core==1.3.2
identify==2.5.31
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
iterative-telemetry==0.0.8
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keybert==0.8.3
keyphrase-vectorizers==0.0.11
kiwisolver==1.4.5
kombu==5.3.4
langcodes==3.3.0
langdetect==1.0.9
llvmlite==0.41.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
myst-parser==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
Pillow==10.1.0
pip-licenses==4.3.3
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
preshed==3.0.9
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
PyDrive2==1.17.0
pygit2==1.13.2
Pygments==2.16.1
pygtrie==2.5.0
pynndescent==0.5.10
pyOpenSSL==23.3.0
pyparsing==3.1.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.11.3
scmrepo==1.4.1
seaborn==0.13.0
semver==3.0.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
spacy==3.7.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.2
Sphinx==5.3.0
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqltrie==0.8.0
srsly==2.4.8
stack-data==0.6.3
stop-words==2018.7.23
sympy==1.12
tabulate==0.9.0
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.24.6
voluptuous==0.14.1
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
yarl==1.9.2
zc.lockfile==3.0.post1
zope.interface==6.1
